Finished new API methods + tidy up
