+++ apps/workpackage-manager/wpm/openproject.py
@@ -22,13 +22,13 @@ class WorkPackageSpec(NamedTuple):
     project_id: int
     description: str = ''
     extra_fields: Dict[str, Any] = {}
-    extra_links: Dict[str, str] = {}
+    extra_custom_options: Dict[str, int] = {}
     status_id: Optional[int] = None
     parent_id: Optional[int] = None
 
     def as_openproject_object(self):
         extra_links = {
-            k: {'href': '/api/v3/' + v} for k, v in self.extra_links.items()
+            k: {'href': '/api/v3/custom_options/{}'.format(v)} for k, v in self.extra_custom_options.items()
         }
         return {
             'subject': self.subject,
